==> pre_processing/01_census_cleanup.ipynb <==
 # Census cleanup 
To obtain the migration matrix, we need census data from 2013 that is compatible with the effective distance data obtained. 
This notebook converts a census dataset into a compatible format to be usable to calculate the migration matrix.

import pandas as pd
import pycountry

population_sizes = pd.read_csv('data_sources/population2013.csv', header=0)
population_sizes.shape


Source: https://data-explorer.oecd.org/vis?tenant=archive&df[ds]=DisseminateArchiveDMZ&df[id]=DF_EDU_DEM&df[ag]=OECD&dq=..&pd=2013%2C2013&to[TIME_PERIOD]=false

Country populations in 2013 (since flight data was from 2013 and so fluxes rely on population structure from 2013)

population_sizes.head()

### Clean

population_sizes = population_sizes[['Country Code', '2014 [YR2014]', 'Country Name']]
population_sizes.reset_index(drop=True, inplace=True)
population_sizes.rename(columns={'Country Code': 'alpha3', '2014 [YR2014]':'population', 'Country Name':'fullname'}, inplace=True)
population_sizes.head()

## Replace alpha3 with alpha2


df = pd.DataFrame(columns=['alpha2', 'population', 'fullname'])
new_index = 0
for index, row in population_sizes.iterrows():
    try: 
        country = pycountry.countries.get(alpha_3 = row.alpha3)
        if country is not None: 
            df.loc[new_index] = [country.alpha_2] + [row.population] + [row.fullname]
            new_index += 1
    except:
        print(f'{row.alpha3} country not found')

df

### Save result

df.to_csv(r'data_sources/census_2013.csv', index=None, sep=',', mode='w')
