==> pre_processing/02_sampling_multipliers.ipynb <==
# Sampling multiplier 
*Important* The data in the sources here is too big to upload to GitHub, the links to the original sources are below, and will have to be placed inside this script if you wish to recalculate sampling multipliers using the same data.

We want a realistic sampling bias in our simulations. This sheet computes sampling multipliers for each country in our simulation by calculating the cases sequenced to number of cases ratio for each country, the most available dataset for this is COVID.

COVID sequencing data is obtained from https://www.biorxiv.org/content/10.1101/2024.07.12.603240v1
COVID case data is obtained from https://ourworldindata.org/covid-cases


import pandas as pd
import pycountry
unclean_seq_m = pd.read_csv('data_sources/Viridian_2M_noShortDel_MAPLE_metaData.tsv', sep='\t')
cases_by_country = pd.read_csv('data_sources/owid-covid-data.csv')
cases_by_country['date'] = pd.to_datetime(cases_by_country['date'], format='%Y-%M-%d')

seq_m = unclean_seq_m[unclean_seq_m['country'].notna()]
seq_m = seq_m[seq_m['country']!= 'UNKNOWN']
seq_m.head()

seq_count = seq_m.groupby('country').size()
def convert(country_name):
    country = pycountry.countries.get(name=country_name)
    if country is None:
        if country_name == 'Russia':
            return 'RUS'
        return country_name
    return country.alpha_3
seq_count.index = seq_count.index.map(lambda x: convert(x))
seq_count.head()

cases_by_country = cases_by_country[cases_by_country['date']<=pd.to_datetime('2022-10-28', format='%Y-%M-%d')]

total_cases_by_country = cases_by_country.groupby('iso_code', as_index=True).aggregate({'new_cases':'sum'})
# Hong Kong is missing, using: https://www.worldometers.info/coronavirus/country/china-hong-kong-sar/

total_cases_by_country.loc['HKG', 'new_cases'] = 2937609.

sampling_proportions = pd.DataFrame(columns=['alpha2','proportion'])
new_index = 0
for country in seq_count.index:
    sampling_prop = seq_count[country]/total_cases_by_country.loc[country]['new_cases']
    sampling_proportions.loc[new_index] = [pycountry.countries.get(alpha_3=country).alpha_2]+[sampling_prop]
    new_index+=1
sampling_proportions['proportion'] = sampling_proportions['proportion'] 
sampling_proportions

sampling_proportions.to_csv('data_sources/sampling_multipliers.csv', index=False)

total_cases_by_country.loc['GBR']
